Update lab1
Split the project into submodules
Fix bags
