+++ protocol/src/main/java/com/github/lexcorp3439/net/lab1/protocol/Utils.java
@@ -1,0 +1,16 @@
+package com.github.lexcorp3439.net.lab1.protocol;
+
+import java.util.Random;
+
+public class Utils {
+    static final String AB = "0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz";
+
+    public static String randomString(int len) {
+        Random rnd = new Random();
+        StringBuilder sb = new StringBuilder(len);
+        for (int i = 0; i < len; i++) {
+            sb.append(AB.charAt(rnd.nextInt(AB.length())));
+        }
+        return sb.toString();
+    }
+}
