--- src/main/java/lab1/Utils.java
@@ -1,16 +1,0 @@
-package lab1;
-
-import java.util.Random;
-
-public class Utils {
-    static final String AB = "0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz";
-
-    public static String randomString(int len) {
-        Random rnd = new Random();
-        StringBuilder sb = new StringBuilder(len);
-        for (int i = 0; i < len; i++) {
-            sb.append(AB.charAt(rnd.nextInt(AB.length())));
-        }
-        return sb.toString();
-    }
-}
